refactor: rewrite Message types between realtime and backend
