+++ backend/src/types/rabbit.ts
@@ -1,13 +1,17 @@
-import type { Message } from "../entities/Message";
-
 export type MessageFromRealtime = {
   content: string;
   createdAt: string;
-  owner: {
-    id: string;
-    name: string; // ?
-  };
+  ownerId: string;
+  campaignId: string;
+};
+
+export type MessageToRealtime = {
+  id: string;
+  content: string;
   campaign: {
+    id: string;
+  };
+  owner: {
     id: string;
   };
 };
@@ -15,7 +19,7 @@ export type MessageFromRealtime = {
 export declare namespace Rabbit {
   type SendMessage = {
     // QueueName: Données envoyées
-    newMessageCallback: Message;
+    newMessageCallback: MessageToRealtime;
   };
 
   type Consume = {
