+++ realtime/src/services/messageBroker.ts
@@ -1,10 +1,10 @@
 import { ioServer } from "..";
 import RabbitMQ from "../lib/RabbitMQ";
-import type { Message, WithoutID } from "../types";
+import type { MessageToBackend } from "../types/message";
 
 const rabbitMQ = RabbitMQ.getInstance(`amqp://${process.env.REALTIME_HOST}`);
 
-export default async function sendToPersist(message: WithoutID<Message>) {
+export default async function sendToPersist(message: MessageToBackend) {
   await rabbitMQ.sendMessage(message, "newMessage");
 }
